# odoo-openapi.yaml
openapi: 3.0.0
info:
  title: "ERP API & Event Handling Service"
  version: 1.0.0
tags:
  - name: Sale Order
    description: Odoo API Contract
servers:
  - url: http://localhost:9000
components:
  parameters:
    authorization:
      in: header
      required: true
      name: Authorization
      description: Bearer <token>
      example: "Bearer XYZ"
      schema:
        type: string
    order_id:
      in: path
      required: true
      name: order_id
      description: Sale Order Id
      example: 36
      schema:
        type: integer
  schemas:
    ClientRequest:
      type: object
      properties:
        name:
          type: string
        partner_id:
          type: integer
        date_order:
          type: string
        company_id:
          type: integer
        order_line:
          type: array
          items:
            type: object
            properties:
              product_id:
                type: integer
              name:
                type: string
              product_uom_qty:
                type: integer
              product_uom:
                type: integer
              price_unit:
                type: integer
  requestBodies:
    ClientRequest:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ClientRequest"
          example:
            name: "EXTERNAL-ORDER-12345"
            partner_id: 11
            date_order: '2021-08-20 08:48:28'
            company_id: 1
            order_line:
              product_id: 6
              product_uom_qty: 1
              product_uom: 1
              price_unit: 4000

paths:
  /order:
    post:
      tags:
        -  Sale Order
      summary: Create Quotation Order
      parameters:
        - $ref: "#/components/parameters/authorization"
      requestBody:
        $ref: "#/components/requestBodies/ClientRequest"
      responses:
        "200":
          description: "Ok"
          content:
            application/json:
              examples:
                response:
                  value:
                    success: true
                    message: "Success"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              examples:
                response:
                  value:
                    success: false
                    message: "Token Not Found"
        "400":
          description: "Bad request"
          content:
            application/json:
              examples:
                response:
                  value:
                    success: false
                    data:
                      name: "Required"
                      partner_id: "Not found"
                      date_order: "Invalid format date"
                      company_id: "Required"
                      order_line: "Is not array"
        "409":
          description: "Conflict"
          content:
            application/json:
              examples:
                response:
                  value:
                    success: false
                    data:
                      name: "Already exists"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              examples:
                response:
                  value:
                    success: false
                    message: "Internal Server Error"

  /order/{order_id}:
    get:
      tags:
        -  Sale Order
      summary: Get Quotation Order
      parameters:
        - $ref: "#/components/parameters/authorization"
        - $ref: "#/components/parameters/order_id"
      responses:
        "200":
          description: "Ok"
          content:
            application/json:
              examples:
                response:
                  value:
                    success: true
                    message: "Data found"
                    data:
                      name: "EXTERNAL-ORDER-12345"
                      partner_id: 11
                      date_order: '2021-08-20 08:48:28'
                      company_id: 1
                      relationship:
                        partner:
                          partner_id: 11
                          name: "Gemini Furniture"
                          address: "317 Fairchild Dr"
                        company:
                          id: 1
                          name: "YourCompany"
                          description: "250 Executive Park Blvd, Suite 3400"
                        product:
                          id: 5
                          name: "Office Chair"
                          description: "FURN_7777"
                          price: 70.0
                        uom:
                            id: 1
                            name: "Units"
                            description: "reference"
                        order_line:
                            product_id: 5
                            price_unit: 400.0
                            product_uom: 1
                            product_uom_qty: 30
        "401":
          description: "Unauthorized"
          content:
            application/json:
              examples:
                response:
                  value:
                    success: false
                    message: "Token Not Found"
        "400":
          description: "Bad request"
          content:
            application/json:
              examples:
                response:
                  value:
                    success: false
                    message: "order id not found"
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              examples:
                response:
                  value:
                    success: false
                    message: "Internal Server Error"
    put:
      tags:
        - Sale Order
      summary: Update Quotation Order
      parameters:
        - $ref: "#/components/parameters/authorization"
        - $ref: "#/components/parameters/order_id"
      requestBody:
        $ref: "#/components/requestBodies/ClientRequest"
      responses:
        "200":
          description: "Ok"
          content:
            application/json:
              examples:
                response:
                  value:
                    success: true
                    message: "Success"
        "401":
          description: "Unauthorized"
          content:
            application/json:
              examples:
                response:
                  value:
                    success: false
                    message: "Token Not Found"
        "400":
          description: "Bad request"
          content:
            application/json:
              examples:
                response:
                  value:
                    success: false
                    data:
                      name: "Required"
                      partner_id: "Not found"
                      date_order: "Invalid format date"
                      company_id: "Required"
                      order_line: "Is not array"
        "409":
          description: "Conflict"
          content:
            application/json:
              examples:
                response:
                  value:
                    success: false
                    data:
                      name: Already exists
        "500":
          description: "Internal Server Error"
          content:
            application/json:
              examples:
                response:
                  value:
                    success: false
                    message: "Internal Server Error"